# contracts/openapi/phase-01.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "browserSession": {
        "in": "cookie",
        "name": "__Host-hdi-session",
        "type": "apiKey"
      },
      "serviceBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hospital Data Intelligence Governance API",
    "version": "0.0.0-phase-01"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/phase-01/consumer-subscriptions": {
      "post": {
        "operationId": "createPhase01ConsumerSubscription",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "minLength": 32,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "governanceObjectId": {
                    "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                    "type": "string"
                  },
                  "projectionSchemaVersion": {
                    "anyOf": [
                      {
                        "enum": [
                          "0"
                        ],
                        "type": "string"
                      },
                      {
                        "enum": [
                          "1"
                        ],
                        "type": "string"
                      }
                    ]
                  },
                  "projectionType": {
                    "enum": [
                      "hdi.price-list"
                    ],
                    "type": "string"
                  },
                  "servicePrincipalId": {
                    "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                    "type": "string"
                  },
                  "subscriptionCode": {
                    "maxLength": 128,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "subscriptionCode",
                  "servicePrincipalId",
                  "governanceObjectId",
                  "projectionType",
                  "projectionSchemaVersion"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "subscriptionId": {
                      "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "subscriptionId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ],
        "summary": "建立精确投影契约的仿真消费者订阅"
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}